requirements.txt
pillow
pytesseract
pdf2image
pandas
spacy
scikit-learn
sqlalchemy
python-dotenv
streamlit
python-multipart
gunicorn
requests
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
python-multipart